json_encode issue
parameters were getting based to javascript as '""' and not allowing new posts to be able to be commented on, disqus error was being returned. Testing locally shows that none of the parameters need to be json_encoded, I presume because we are not building a json object to be passed.